+++ src/Oro/Bundle/WebsiteSearchSuggestionBundle/Provider/SuggestionProvider.php
@@ -11,6 +11,8 @@ use Oro\Bundle\WebsiteSearchSuggestionBundle\Splitter\PhraseSplitter;
  */
 class SuggestionProvider
 {
+    private ?int $phraseChunkSize = null;
+
     public function __construct(
         private ProductsProvider $productsProvider,
         private PhraseSplitter   $phraseSplitter,
@@ -18,32 +20,42 @@ class SuggestionProvider
     ) {
     }
 
-    public function getLocalizedSuggestionPhrasesGroupedByProductId(array $productIds): array
+    public function setChunkSize(int $chunkSize): void
     {
-        $productsByLocalizedPhrase = [];
+        $this->phraseChunkSize = $chunkSize;
+    }
 
+    public function getLocalizedSuggestionPhrasesGroupedByProductId(array $productIds): \Generator
+    {
         $productsSkuNames = $this->productsProvider->getProductsSkuAndNames($productIds);
-
         $localizations = $this->localizationHelper->getLocalizations();
 
-        foreach ($productsSkuNames as $productId => $textGroup) {
-            foreach ($localizations as $localization) {
+        foreach ($localizations as $localization) {
+            $result = [];
+            foreach ($productsSkuNames as $productId => $textGroup) {
                 $productName = $this->localizationHelper->getLocalizedValue(
                     new ArrayCollection($textGroup['names']),
                     $localization
                 );
 
-                $phrases = array_merge(
+                $phrases = \array_merge(
                     $this->phraseSplitter->split($textGroup['sku']),
                     $this->phraseSplitter->split($productName->getString())
                 );
 
                 foreach ($phrases as $phrase) {
-                    $productsByLocalizedPhrase[$localization->getId()][$phrase][] = $productId;
+                    $result[$phrase][$productId] = $productId;
+                }
+
+                if (\count($result) > $this->phraseChunkSize) {
+                    yield $localization->getId() => $result;
+                    $result = [];
                 }
             }
-        }
 
-        return $productsByLocalizedPhrase;
+            if (!empty($result)) {
+                yield $localization->getId() => $result;
+            }
+        }
     }
 }